+++ app-main/src/main/java/net/causw/app/main/domain/user/account/service/UserAdminService.java
@@ -25,6 +25,7 @@ import net.causw.app.main.domain.user.account.service.dto.response.UserRestoreWi
 import net.causw.app.main.domain.user.account.service.dto.response.UserRoleUpdateResult;
 import net.causw.app.main.domain.user.account.service.dto.result.DeletedUserListItemDto;
 import net.causw.app.main.domain.user.account.service.implementation.DroppedUserIdentifierWriter;
+import net.causw.app.main.domain.user.account.service.implementation.UserAccountCleanupWriter;
 import net.causw.app.main.domain.user.account.service.implementation.UserAdminActionLogWriter;
 import net.causw.app.main.domain.user.account.service.implementation.UserReader;
 import net.causw.app.main.domain.user.account.service.implementation.UserWriter;
@@ -44,7 +45,7 @@ public class UserAdminService {
 	private final UserProfileImageReader userProfileImageReader;
 	private final DroppedUserIdentifierWriter droppedUserIdentifierWriter;
 	private final UserAccountService userAccountService;
-	private final UserAccountCleanupService userAccountCleanupService;
+	private final UserAccountCleanupWriter userAccountCleanupWriter;
 
 	// 필터링 조건과 페이징 정보를 기반으로 전체 사용자 목록 조회
 	@Transactional(readOnly = true)
@@ -75,7 +76,7 @@ public class UserAdminService {
 		UserState beforeState = targetUser.getState();
 		Set<Role> beforeRoles = new HashSet<>(targetUser.getRoles());
 
-		userAccountCleanupService.cleanupForDrop(targetUser);
+		userAccountCleanupWriter.cleanupForDrop(targetUser);
 
 		lockerReader.findByUserId(targetUser.getId()).ifPresent(locker -> {
 			lockerWriter.releaseLocker(locker, adminUser, targetUser.getEmail(), targetUser.getName());
